Updated the makefile

+++ Makefile
@@ -15,7 +15,7 @@ test: lint generate-tests
 	$(BIN)/mocha $(MOCHA_OPTS) --reporter $(REPORTER) $(TEST_FILES)
 
 generate-tests:
-	rm $(GENERATED_TEST_FILES) || true
+	[ -f $(GENERATED_TEST_FILES) ] && rm $(GENERATED_TEST_FILES) || true
 	@node util/generate-mocha-test-cases.js
 
 test-reports: lib-cov
